[Feat] Participant 레포지토리, 서비스, DTO(미완)

+++ src/main/java/com/shinhan/soloplay/raid/RaidEntity.java
@@ -22,7 +22,7 @@ public class RaidEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer raidId;
+	private Long raidId;
 	
 	private int hitPoint;
 	private Timestamp startTime;
